+++ packages/faceted-search/src/components/Badges/BadgeSlider/BadgeSlider.component.js
@@ -41,7 +41,8 @@ const BadgeSlider = ({
 }) => {
 	const currentOperators = useMemo(() => operators || createDefaultOperators(t), [operators, t]);
 	const currentOperator = operator || currentOperators[0];
-	const labelValue = (value || defaultValue) != null && `${value || defaultValue}${unit}`;
+	const numericValue = value != null ? value : defaultValue;
+	const labelValue = numericValue != null && `${numericValue}${unit}`;
 	const badgeTextId = `${id}-badge-slider`;
 
 	return (
@@ -65,6 +66,7 @@ const BadgeSlider = ({
 					onSubmit={onSubmitBadge}
 					operator={currentOperator}
 					value={badgeValue}
+					defaultValue={defaultValue}
 					feature={category || label}
 					unit={unit}
 					t={t}